Subgroup/ LabelRankingMatrix: fix redundant cast compiler warning.
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2484 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/LabelRankingMatrix.java
@@ -171,11 +171,12 @@ public class LabelRankingMatrix
 			float theLabelMean2 = labelMean(itsMatrix, i);
 			for (int j=0; j<itsSize; j++)
 			{
-				if (i!=j)	aDistance += (theMatrix.itsMatrix[i][j] - theLabelMean)*(itsMatrix[i][j] - theLabelMean2);
+				if (i!=j)
+					aDistance += (theMatrix.itsMatrix[i][j] - theLabelMean)*(itsMatrix[i][j] - theLabelMean2);
 			}
 			aCovariance += -aDistance/(itsSize-2);
 		}
-		return (float) aCovariance/itsSize;
+		return aCovariance/itsSize;
 	}
 	
 	public float labelMean(float[][] theRankingMatrix, int label)
